refactor: centralize admin user helpers
